partial backup support.

+++ backup.sh
@@ -1,18 +1,21 @@
-#!/usr/bin/env bash
+#!/usr/bin/env fish
 
-rsync -a --exclude='logs' ~/.weechat .
-tar czf weechat.tar.gz .weechat
-rm weechat.tar.gz.gpg
-gpg --recipient amosbird@gmail.com --encrypt weechat.tar.gz
-rm weechat.tar.gz
+if test (count $argv) != 0
+  exit 1
+  rsync -a --exclude='logs' ~/.weechat .
+  tar czf weechat.tar.gz .weechat
+  rm weechat.tar.gz.gpg
+  gpg --recipient amosbird@gmail.com --encrypt weechat.tar.gz
+  rm weechat.tar.gz
 
-rsync -a --exclude='Cache' --exclude='Media Cache' --exclude='Local Storage' ~/.config/vivaldi .
-tar czf vivaldi.tar.gz vivaldi
-rm vivaldi.tar.gz.gpg
-gpg --recipient amosbird@gmail.com --encrypt vivaldi.tar.gz
-rm vivaldi.tar.gz
-rm -rf vivaldi
-rm -rf .weechat
+  rsync -a --exclude='Cache' --exclude='Media Cache' --exclude='Local Storage' ~/.config/vivaldi .
+  tar czf vivaldi.tar.gz vivaldi
+  rm vivaldi.tar.gz.gpg
+  gpg --recipient amosbird@gmail.com --encrypt vivaldi.tar.gz
+  rm vivaldi.tar.gz
+  rm -rf vivaldi
+  rm -rf .weechat
+end
 
 cp ~/.spacemacs spacemacs
 rsync -a ~/.emacs.d/private/ spacemacs/
